chore(m0): move ESC/I sources into src/esci/ subfolder
Move legacy ESC/I protocol support files (formerly prefixed esci-legacy-) into
a new src/esci/ subdirectory alongside test support files:
- src/esci-legacy.ts → src/esci/commands.ts
- src/esci-legacy.test.ts → src/esci/commands.test.ts
- src/esci-legacy-luts.ts → src/esci/luts.ts
- src/raw-to-jpeg.ts → src/esci/raw-to-jpeg.ts
- src/raw-to-jpeg.test.ts → src/esci/raw-to-jpeg.test.ts
- src/scanner-legacy.ts → src/esci/scanner.ts
- src/scanner-legacy.test.ts → src/esci/scanner.test.ts
- src/scanner-legacy-diagnose.test.ts → src/esci/scanner-diagnose.test.ts
- src/test-support/legacy-replay.ts → src/esci/test-support/replay.ts
- src/test-support/fake-tcp-socket.ts → src/esci/test-support/fake-tcp-socket.ts

Update all imports in moved and consuming files. Update tsconfig.json exclude
pattern to preserve test-support file exclusion across the new subdirectory.

Test count: 294 passed + 1 skipped (unchanged).
