+++ src/esci/raw-to-jpeg.test.ts
